Move runStatusCmd in helpers_tests.go
